first GET routes

See merge request informatik/qe/swess23/group4/g4t3!19

+++ src/main/java/at/qe/skeleton/repositories/MeasurementRepository.java
@@ -4,5 +4,5 @@ import at.qe.skeleton.model.Measurement;
 
 public interface MeasurementRepository extends AbstractRepository<Measurement, Long> {
 
-    Measurement findFirstById(Long id);
+    Measurement findFirstById(Integer id);
 }
